requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.39.0
requests==2.32.3
openai==1.52.2
